# packages/api-contract/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nahhasio local client API",
    "version": "1.0.0",
    "description": "Initial local owner API. Ordinary family accounts require policy enforcement and are not accepted yet. Playback/download endpoints are not implemented."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/works": {
      "get": {
        "operationId": "listWorks",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Owner access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100000,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 24
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["title", "year-desc", "year-asc", "updated-desc", "added-desc"],
              "default": "title"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["animated", "live-action"]
            }
          },
          {
            "name": "audience",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["general", "teen", "young-adult", "adult"]
            }
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["announced", "airing", "completed", "unknown"]
            }
          },
          {
            "name": "yearFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9999
            }
          },
          {
            "name": "yearTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9999
            }
          },
          {
            "name": "includePrivate",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Owner opt-in; defaults false."
          },
          {
            "name": "planet",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "description": "Owner-only catalog. All filters combine with AND; genre matches its vocabulary slug. Text search matches title, Arabic title and aliases. Stable SQL pagination."
      }
    },
    "/api/v1/works/{id}": {
      "get": {
        "operationId": "getWork",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Owner access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/v1/catalog/filters": {
      "get": {
        "operationId": "getCatalogFilters",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogFilters"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Owner access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/artwork/{id}": {
      "get": {
        "operationId": "getArtwork",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registered artwork. Send Bearer authorization. URL is relative to the configured server.",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/gif": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Owner access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/catalog/home": {
      "get": {
        "operationId": "getHomeFeed",
        "responses": {
          "200": {
            "description": "Bounded public family activity and anticipated installments, nearest known date first. Published comments/reviews from active discoverable accounts only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeFeed"
                }
              }
            }
          },
          "401": {
            "description": "Owner authentication required"
          }
        }
      }
    },
    "/api/v1/catalog/browse": {
      "get": {
        "operationId": "browseCatalog",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "view",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["works", "installments"]
            }
          },
          {
            "name": "privacy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["public", "all", "private"]
            }
          },
          {
            "name": "filters",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "yearFrom",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "yearTo",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated owner viewing API; bounded read-only result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter values"
          },
          "401": {
            "description": "Owner authentication required"
          }
        }
      }
    },
    "/api/v1/catalog/facets": {
      "get": {
        "operationId": "getCatalogFacets",
        "parameters": [
          {
            "name": "view",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["works", "installments"]
            }
          },
          {
            "name": "privacy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["public", "all", "private"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated owner viewing API; bounded read-only result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FacetCatalog"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter values"
          },
          "401": {
            "description": "Owner authentication required"
          }
        }
      }
    },
    "/api/v1/catalog/recommendations": {
      "get": {
        "operationId": "getRecommendations",
        "parameters": [
          {
            "name": "workId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated owner viewing API; bounded read-only result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter values"
          },
          "401": {
            "description": "Owner authentication required"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["message"],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "Artwork": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "url",
          "mimeType",
          "width",
          "height",
          "byteSize",
          "sha256",
          "originalFilename",
          "focalX",
          "focalY"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "width": {
            "type": "integer"
          },
          "height": {
            "type": "integer"
          },
          "byteSize": {
            "type": "integer"
          },
          "sha256": {
            "type": "string"
          },
          "originalFilename": {
            "type": "string"
          },
          "focalX": {
            "type": "integer"
          },
          "focalY": {
            "type": "integer"
          }
        }
      },
...
      "WorkDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "canonicalTitle",
          "titleAr",
          "summary",
          "releaseYear",
          "format",
          "audience",
          "age",
          "poster",
          "installmentCount",
          "episodeCount",
          "contentWarnings",
          "analysisNotes",
          "sexualityRisk",
          "behavioralRisk",
          "theologyRisk",
          "externalIds",
          "aliases",
          "trivia",
          "genres",
          "contributions",
          "relations",
          "artwork",
          "installments",
          "curatorNotes",
          "qualityScore",
          "verifiedAt",
          "createdAt",
          "updatedAt",
          "externalReferences",
          "tones",
          "tags",
          "countries",
          "planets",
          "awards",
          "sortTitle",
          "isPrivate",
          "score",
          "banner",
          "logo"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "canonicalTitle": {
            "type": "string"
          },
          "titleAr": {
            "type": "string",
            "nullable": true
          },
          "summary": {
            "type": "string"
          },
          "releaseYear": {
            "type": "integer",
            "nullable": true
          },
          "format": {
            "type": "string",
            "enum": ["animated", "live-action"]
          },
          "audience": {
            "type": "string",
            "enum": ["general", "teen", "young-adult", "adult"]
          },
          "age": {
            "type": "string"
          },
          "poster": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Artwork"
              }
            ],
            "nullable": true
          },
          "installmentCount": {
            "type": "integer"
          },
          "episodeCount": {
            "type": "integer"
          },
          "contentWarnings": {
            "type": "string",
            "nullable": true
          },
          "analysisNotes": {
            "type": "string",
            "nullable": true
          },
          "sexualityRisk": {
            "type": "string",
            "enum": ["none", "low", "medium", "high"]
          },
          "behavioralRisk": {
            "type": "string",
            "enum": ["none", "low", "medium", "high"]
          },
          "theologyRisk": {
            "type": "string",
            "enum": ["none", "low", "medium", "high"]
          },
          "externalIds": {
            "$ref": "#/components/schemas/ExternalIds"
          },
          "aliases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Alias"
            }
          },
          "trivia": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vocabulary"
            }
          },
          "contributions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Contribution"
            }
          },
          "relations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Relation"
            }
          },
          "artwork": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtworkAssignment"
            }
          },
          "installments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Installment"
            }
          },
          "curatorNotes": {
            "type": "string"
          },
          "qualityScore": {
            "type": "integer"
          },
          "verifiedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "externalReferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExternalReference"
            }
          },
          "tones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vocabulary"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vocabulary"
            }
          },
          "countries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vocabulary"
            }
          },
          "planets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Planet"
            }
          },
          "awards": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AwardRecognition"
            }
          },
          "sortTitle": {
            "type": "string"
          },
          "isPrivate": {
            "type": "boolean"
          },
          "score": {
            "$ref": "#/components/schemas/ScoreSummary"
          },
          "banner": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Artwork"
              }
            ],
            "nullable": true
          },
          "logo": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Artwork"
              }
            ],
            "nullable": true
          }
        }
      },
      "WorkPage": {
        "type": "object",
        "additionalProperties": false,
        "required": ["items", "page", "pageSize", "total"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkSummary"
            }
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "CatalogFilters": {
        "type": "object",
        "additionalProperties": false,
        "required": ["genres", "formats", "audiences", "statuses", "yearMin", "yearMax", "planets"],
        "properties": {
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vocabulary"
            }
          },
          "formats": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "audiences": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "statuses": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "yearMin": {
            "type": "integer",
            "nullable": true
          },
          "yearMax": {
            "type": "integer",
            "nullable": true
          },
          "planets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogPlanet"
            }
          }
        }
      },
...
      "HomeFeed": {
        "type": "object",
        "required": ["comments", "upcoming"],
        "properties": {
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FamilyActivity"
            }
          },
          "upcoming": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpcomingInstallment"
            }
          }
        }
      },
...
      "CatalogPage": {
        "type": "object",
        "required": ["items", "total", "page", "pageSize"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogEntry"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          }
        }
      },
...
      "FacetCatalog": {
        "type": "object",
        "required": ["groups", "yearMin", "yearMax"],
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FacetGroup"
            }
          },
          "yearMin": {
            "type": "integer",
            "nullable": true
          },
          "yearMax": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
      "RecommendationPage": {
        "type": "object",
        "required": ["items", "basis"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkSummary"
            }
          },
          "basis": {
            "type": "string",
            "enum": ["related", "personal", "editorial"]
          }
        }
      }
    }
  }
}